requirements.txt
Django>=5.0
djangorestframework
drf-yasg
requests
beautifulsoup4
psycopg2-binary
python-dotenv
gunicorn
